# swagger.yaml
openapi: 3.0.0
info:
  title: 'Oetker code API challenge'
  version: 1.0.0
servers:
  -
    url: 'localhost:8000/api/v1'
    description: Development server
paths:
  '/animals/requests/{request_id}':
    delete:
      description: 'remove adaptation request'
      operationId: 'App\Http\Controllers\Api\V1\AdaptationRequestsController::delete'
      parameters:
        -
          name: request_id
          in: path
          description: 'ID of request'
          required: true
          schema:
            type: integer
      responses:
        '404':
          description: 'Request not found'
        '204':
          description: 'Delete request'
  /animals:
    get:
      description: 'Get animals by params'
      operationId: 'App\Http\Controllers\Api\V1\AnimalsController::getAnimals'
      parameters:
        -
          name: shelter_id
          in: query
          description: 'Shelter id'
          schema:
            type: integer
        -
          name: name
          in: query
          description: 'Name of animal'
          schema:
            type: string
        -
          name: status
          in: query
          description: 'Status of animal'
          schema:
            type: string
        -
          name: offset
          in: query
          description: Offset
          schema:
            type: integer
        -
          name: limit
          in: query
          description: Limit
          schema:
            type: integer
      responses:
        '404':
          description: 'Animal not found'
        '200':
          description: OK
    post:
      description: 'Add animal'
      operationId: 'App\Http\Controllers\Api\V1\AnimalsController::create'
      parameters:
        -
          name: name
          in: query
          description: 'Animal name'
          required: true
          schema:
            type: string
        -
          name: shelter_id
          in: query
          description: 'Shelter id associated with animal'
          required: true
          schema:
            type: integer
        -
          name: race
          in: query
          description: 'Race of animal'
          required: true
          schema:
            type: string
        -
          name: picture
          in: query
          description: 'Picture of animal'
          schema:
            type: string
        -
          name: medical_condition
          in: query
          description: 'Mecial condition of animal'
          schema:
            type: string
        -
          name: status
          in: query
          description: 'Status of animal shoould be adaptable, pending or adapted'
          schema:
            type: string
      responses:
        '201':
          description: 'Animal created'
        '422':
          description: 'Validation failed'
  '/animals/{animal_id}/adapted':
    put:
      description: 'Set status animal to adapteble'
      operationId: 'App\Http\Controllers\Api\V1\AnimalsController::adaptable'
      parameters:
        -
          name: animal_id
          in: path
          description: 'ID of animal to delete'
          required: true
          schema:
            type: integer
      responses:
        '404':
          description: 'Animal not found'
        '200':
          description: 'Status was update'
  '/animals/{animal_id}':
    delete:
      description: 'remove animal by id'
      operationId: 'App\Http\Controllers\Api\V1\AnimalsController::delete'
      parameters:
        -
          name: animal_id
          in: path
          description: 'ID of animal to delete'
          required: true
          schema:
            type: integer
      responses:
        '404':
          description: 'Animal not found'
        '204':
          description: 'Delete animal'
  /shelters:
    get:
      description: 'Get shelters lists'
      operationId: 'App\Http\Controllers\Api\V1\SheltersController::getShelters'
      parameters:
        -
          name: limit
          in: query
          description: 'Limit default 100'
          schema:
            type: integer
        -
          name: offset
          in: query
          description: Offset
          schema:
            type: integer
      responses:
        '200':
          description: OK
    post:
      description: 'Add shelter'
      operationId: 'App\Http\Controllers\Api\V1\SheltersController::create'
      parameters:
        -
          name: name
          in: query
          description: 'Shelter name'
          required: true
          schema:
            type: string
        -
          name: address
          in: query
          description: 'Shelter address'
          required: true
          schema:
            type: string
      responses:
        '201':
          description: 'Shelter created'
        '422':
          description: 'Validation failed'
  '/shelters/{shelter_id}':
    delete:
      description: 'remove shelter by id'
      operationId: 'App\Http\Controllers\Api\V1\SheltersController::delete'
      parameters:
        -
          name: shelter_id
          in: path
          description: 'ID of shelter to delete'
          required: true
          schema:
            type: integer
      responses:
        '404':
          description: 'Shelter not found'
        '204':
          description: 'Delete shelter'
  '/shelters/{shelter_id}/workers':
    post:
      description: 'Add worker'
      operationId: 'App\Http\Controllers\Api\V1\WorkersController::create'
      parameters:
        -
          name: name
          in: query
          description: 'Worker name'
          required: true
          schema:
            type: string
        -
          name: shelter_id
          in: path
          description: 'Shelter id'
          required: true
          schema:
            type: integer
      responses:
        '201':
          description: 'Shelter created'
        '422':
          description: 'Validation failed'
  '/shelters/workers/{worker_id}':
    delete:
      description: 'remove worker by id'
      operationId: 'App\Http\Controllers\Api\V1\WorkersController::delete'
      parameters:
        -
          name: worker_id
          in: path
          description: 'ID of worker to delete'
          required: true
          schema:
            type: integer
      responses:
        '404':
          description: 'Worker not found'
        '204':
          description: 'Delete worker'